test: reconcile characterization gate shape
